feat: convert AcceptInvitation to a form request
