+++ database/factories/InvitationFactory.php
@@ -20,7 +20,7 @@ class InvitationFactory extends Factory
      *
      * @return array
      */
-    public function definition()
+    public function definition(): array
     {
         $roles = config('hearth.organizations.roles');
